requirements.txt
# Image processing
numpy==1.16.6
scipy==1.4.1
pillow
opencv-python
imageio==2.27.0

# Device connection
adbutils==0.11.0
uiautomator2==2.16.17
uiautomator2cache==0.3.0.1
wrapt==1.13.1
retrying
lz4
av==10.0.0
psutil==5.9.3

# Utils
rich==11.2.0
tqdm
jellyfish==0.11.2
pyyaml
inflection
pydantic
aiofiles
prettytable==2.2.1
anyio==1.3.1

# Pushing
onepush==1.4.0
pycryptodome==3.9.9
pypresence==4.2.1

# Ocr
cnocr==1.2.2
mxnet==1.6.0

# Webui
pywebio==1.6.2
starlette==0.14.2
uvicorn[standard]==0.17.6
alas-webapp==0.3.7
zerorpc==0.6.3
pyzmq==22.3.0
